Mejora script con verificación de archivo

+++ respaldos/script_respaldo.sh
@@ -5,6 +5,11 @@ mkdir -p respaldos
 fecha=$(date +%F)
 archivo_respaldo="respaldos/informe_$fecha.txt"
 
-cp docs/informe.txt "$archivo_respaldo"
+if [ -f docs/informe.txt ]; then
+    cp docs/informe.txt "$archivo_respaldo"
 
 echo "Respaldo creado en: $archivo_respaldo"
+
+else
+    echo "Error: El archivo informe.txt no existe en la carpeta docs."
+fi
